+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.29.2]
+## [0.29.2] - 2024/11/29
 
 ### Fixed
 
@@ -13,7 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   This bug was introduced in version 0.28.
   [#67](https://github.com/pyodide/pyodide-build/pull/67)
 
-## [0.29.1]
+## [0.29.1] - 2024/11/27
 
 ### Added